fix for desktop build with int32 times

+++ tulip/shared/alles.c
@@ -232,13 +232,13 @@ void alles_init_multicast(uint8_t local_node) {
 extern uint8_t ipv4_quartet;
 extern char githash[8];
 int16_t client_id;
-int64_t clocks[255];
-int64_t ping_times[255];
+int32_t clocks[255];
+int32_t ping_times[255];
 uint8_t alive = 1;
 
-extern int64_t computed_delta ; // can be negative no prob, but usually host is larger # than client
+extern int32_t computed_delta ; // can be negative no prob, but usually host is larger # than client
 extern uint8_t computed_delta_set ; // have we set a delta yet?
-extern int64_t last_ping_time;
+extern int32_t last_ping_time;
 
 amy_err_t sync_init() {
     client_id = -1; // for now
@@ -252,7 +252,7 @@ amy_err_t sync_init() {
 void alles_parse_message(char *message, uint16_t length) {
     uint8_t mode = 0;
     int16_t client = -1;
-    int64_t sync = -1;
+    int32_t sync = -1;
     int8_t sync_index = -1;
     uint8_t ipv4 = 0;
     uint16_t start = 0;
@@ -316,13 +316,13 @@ void alles_parse_message(char *message, uint16_t length) {
     }
 }
 
-void update_map(uint8_t client, uint8_t ipv4, int64_t time) {
+void update_map(uint8_t client, uint8_t ipv4, int32_t time) {
     // I'm called when I get a sync response or a regular ping packet
     // I update a map of booted devices.
 
     //fprintf(stderr,"[%d %d] Got a sync response client %d ipv4 %d time %lld\n",  ipv4_quartet, client_id, client , ipv4, time);
     clocks[ipv4] = time;
-    int64_t my_sysclock = amy_sysclock();
+    int32_t my_sysclock = amy_sysclock();
     ping_times[ipv4] = my_sysclock;
 
     // Now I basically see what index I would be in the list of booted synths (clocks[i] > 0)
@@ -344,7 +344,7 @@ void update_map(uint8_t client, uint8_t ipv4, int64_t time) {
             // If this is not me....
             if(i != ipv4_quartet) {
                 // predicted time is what we think the alive node should be at by now
-                int64_t predicted_time = (my_sysclock - ping_times[i]) + clocks[i];
+                int32_t predicted_time = (my_sysclock - ping_times[i]) + clocks[i];
                 if(my_sysclock >= predicted_time) my_new_client_id--;
             } else {
                 my_new_client_id--;
@@ -360,9 +360,9 @@ void update_map(uint8_t client, uint8_t ipv4, int64_t time) {
     }
 }
 
-void handle_sync(int64_t time, int8_t index) {
+void handle_sync(int32_t time, int8_t index) {
     // I am called when I get an s message, which comes along with host time and index
-    int64_t sysclock = amy_sysclock();
+    int32_t sysclock = amy_sysclock();
     char message[100];
     // Before I send, i want to update the map locally
     //fprintf(stderr, "handle_sync %d %d\n", client_id, ipv4_quartet);
@@ -377,7 +377,7 @@ void handle_sync(int64_t time, int8_t index) {
     //if(old_cd != computed_delta) printf("Changed computed_delta from %lld to %lld on sync\n", old_cd, computed_delta);
 }
 
-void ping(int64_t sysclock) {
+void ping(int32_t sysclock) {
     char message[100];
     //printf("[%d %d] pinging with %lld\n", ipv4_quartet, client_id, sysclock);
     sprintf(message, "_s%" PRIi64 "i-1c%dr%dy%dZ", sysclock, client_id, ipv4_quartet, 0);
